requirements/base.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
baron==0.9
black==19.10b0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
dj-email-url==1.0.0
Django==3.0.6
django-cache-url==3.1.2
django-environ==0.4.5
django-environs==0.1.1
django-filter==2.3.0
django-model-utils==3.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.1.0
idna==2.9
marshmallow==3.6.0
pathspec==0.8.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redbaron==0.9.2
regex==2020.5.14
requests==2.23.0
rply==0.7.7
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9

